LayoutTests/platform/chromium/fast/forms/calendar-picker/week-picker-key-operations.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../../../fast/js/resources/js-test-pre.js"></script>
<script src="../../../../../fast/forms/resources/picker-common.js"></script>
<script src="resources/calendar-picker-common.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<input type=week id=week value="2000-W01">
<script>
description("Tests if month picker key bindings work as expected.");

debug('Check that page popup doesn\'t exist at first.');
shouldBeNull('document.getElementById("mock-page-popup")');

openPicker(document.getElementById('week'), test1);

function test1() {
    debug('Check that page popup exists.');
    shouldBe('popupWindow.pagePopupController.toString()', '"[object PagePopupController]"');
    
    shouldBeEqualToString('currentMonth()', '2000-01');
    shouldBeEqualToString('selectedValue()', '2000-W01');
    shouldBeEqualToString('selectedDayCells()', '2000-01-03,2000-01-04,2000-01-05,2000-01-06,2000-01-07,2000-01-08,2000-01-09');
    shouldBeEqualToString('highlightedValue()', '2000-W01');
    shouldBeEqualToString('highlightedDayCells()', '2000-01-03,2000-01-04,2000-01-05,2000-01-06,2000-01-07,2000-01-08,2000-01-09');

    debug('Check that arrow keys work properly.');
    
    debug("press up");
    eventSender.keyDown('upArrow');
    shouldBeEqualToString('currentMonth()', '2000-01');
    shouldBeEqualToString('selectedValue()', '2000-W01');
    shouldBeEqualToString('selectedDayCells()', '2000-01-03,2000-01-04,2000-01-05,2000-01-06,2000-01-07,2000-01-08,2000-01-09');
    shouldBeEqualToString('highlightedValue()', '1999-W52');
    shouldBeEqualToString('highlightedDayCells()', '1999-12-27,1999-12-28,1999-12-29,1999-12-30,1999-12-31,2000-01-01,2000-01-02');
    
    debug("press up");
    eventSender.keyDown('upArrow');
    shouldBeEqualToString('currentMonth()', '1999-12');
    shouldBeEqualToString('selectedValue()', '2000-W01');
    shouldBeEqualToString('selectedDayCells()', '');
    shouldBeEqualToString('highlightedValue()', '1999-W51');
    shouldBeEqualToString('highlightedDayCells()', '1999-12-20,1999-12-21,1999-12-22,1999-12-23,1999-12-24,1999-12-25,1999-12-26');
    
    debug("press down");
    eventSender.keyDown('downArrow');
    shouldBeEqualToString('currentMonth()', '1999-12');
    shouldBeEqualToString('selectedValue()', '2000-W01');
    shouldBeEqualToString('selectedDayCells()', '');
    shouldBeEqualToString('highlightedValue()', '1999-W52');
    shouldBeEqualToString('highlightedDayCells()', '1999-12-27,1999-12-28,1999-12-29,1999-12-30,1999-12-31,2000-01-01');
    
    debug("press down");
    eventSender.keyDown('downArrow');
    shouldBeEqualToString('currentMonth()', '2000-01');
    shouldBeEqualToString('selectedValue()', '2000-W01');
    shouldBeEqualToString('selectedDayCells()', '2000-01-03,2000-01-04,2000-01-05,2000-01-06,2000-01-07,2000-01-08,2000-01-09');
    shouldBeEqualToString('highlightedValue()', '2000-W01');
    shouldBeEqualToString('highlightedDayCells()', '2000-01-03,2000-01-04,2000-01-05,2000-01-06,2000-01-07,2000-01-08,2000-01-09');

    debug("press down");
    eventSender.keyDown('downArrow');
    shouldBeEqualToString('currentMonth()', '2000-01');
    shouldBeEqualToString('selectedValue()', '2000-W01');
    shouldBeEqualToString('selectedDayCells()', '2000-01-03,2000-01-04,2000-01-05,2000-01-06,2000-01-07,2000-01-08,2000-01-09');
    shouldBeEqualToString('highlightedValue()', '2000-W02');
    shouldBeEqualToString('highlightedDayCells()', '2000-01-10,2000-01-11,2000-01-12,2000-01-13,2000-01-14,2000-01-15,2000-01-16');
    
    debug("press left");
    eventSender.keyDown('leftArrow');
    shouldBeEqualToString('currentMonth()', '2000-01');
    shouldBeEqualToString('selectedValue()', '2000-W01');
    shouldBeEqualToString('selectedDayCells()', '2000-01-03,2000-01-04,2000-01-05,2000-01-06,2000-01-07,2000-01-08,2000-01-09');
    shouldBeEqualToString('highlightedValue()', '2000-W01');
    shouldBeEqualToString('highlightedDayCells()', '2000-01-03,2000-01-04,2000-01-05,2000-01-06,2000-01-07,2000-01-08,2000-01-09');

    debug("press right");
    eventSender.keyDown('rightArrow');
    shouldBeEqualToString('currentMonth()', '2000-01');
    shouldBeEqualToString('selectedValue()', '2000-W01');
    shouldBeEqualToString('selectedDayCells()', '2000-01-03,2000-01-04,2000-01-05,2000-01-06,2000-01-07,2000-01-08,2000-01-09');
    shouldBeEqualToString('highlightedValue()', '2000-W02');
    shouldBeEqualToString('highlightedDayCells()', '2000-01-10,2000-01-11,2000-01-12,2000-01-13,2000-01-14,2000-01-15,2000-01-16');

    debug("press m");
    eventSender.keyDown('m');
    shouldBeEqualToString('currentMonth()', '2000-02');
    shouldBeEqualToString('selectedValue()', '2000-W01');
    shouldBeEqualToString('selectedDayCells()', '');
    shouldBeEqualToString('highlightedValue()', '2000-W07');
    shouldBeEqualToString('highlightedDayCells()', '2000-02-14,2000-02-15,2000-02-16,2000-02-17,2000-02-18,2000-02-19,2000-02-20');

    debug("press shift + m");
    eventSender.keyDown('M');
    shouldBeEqualToString('currentMonth()', '2000-01');
    shouldBeEqualToString('selectedValue()', '2000-W01');
    shouldBeEqualToString('selectedDayCells()', '2000-01-03,2000-01-04,2000-01-05,2000-01-06,2000-01-07,2000-01-08,2000-01-09');
    shouldBeEqualToString('highlightedValue()', '2000-W02');
    shouldBeEqualToString('highlightedDayCells()', '2000-01-10,2000-01-11,2000-01-12,2000-01-13,2000-01-14,2000-01-15,2000-01-16');

    debug("press y");
    eventSender.keyDown('y');
    shouldBeEqualToString('currentMonth()', '2001-01');
    shouldBeEqualToString('selectedValue()', '2000-W01');
    shouldBeEqualToString('selectedDayCells()', '');
    shouldBeEqualToString('highlightedValue()', '2001-W03');
    shouldBeEqualToString('highlightedDayCells()', '2001-01-15,2001-01-16,2001-01-17,2001-01-18,2001-01-19,2001-01-20,2001-01-21');

    debug("press shift + y");
    eventSender.keyDown('Y');
    shouldBeEqualToString('currentMonth()', '2000-01');
    shouldBeEqualToString('selectedValue()', '2000-W01');
    shouldBeEqualToString('selectedDayCells()', '2000-01-03,2000-01-04,2000-01-05,2000-01-06,2000-01-07,2000-01-08,2000-01-09');
    shouldBeEqualToString('highlightedValue()', '2000-W02');
    shouldBeEqualToString('highlightedDayCells()', '2000-01-10,2000-01-11,2000-01-12,2000-01-13,2000-01-14,2000-01-15,2000-01-16');

    debug("press d");
    eventSender.keyDown('d');
    shouldBeEqualToString('currentMonth()', '2010-01');
    shouldBeEqualToString('selectedValue()', '2000-W01');
    shouldBeEqualToString('selectedDayCells()', '');
    shouldBeEqualToString('highlightedValue()', '2010-W02');
    shouldBeEqualToString('highlightedDayCells()', '2010-01-11,2010-01-12,2010-01-13,2010-01-14,2010-01-15,2010-01-16,2010-01-17');

    debug("press shift + d");
    eventSender.keyDown('D');
    shouldBeEqualToString('currentMonth()', '2000-01');
    shouldBeEqualToString('selectedValue()', '2000-W01');
    shouldBeEqualToString('selectedDayCells()', '2000-01-03,2000-01-04,2000-01-05,2000-01-06,2000-01-07,2000-01-08,2000-01-09');
    shouldBeEqualToString('highlightedValue()', '2000-W02');
    shouldBeEqualToString('highlightedDayCells()', '2000-01-10,2000-01-11,2000-01-12,2000-01-13,2000-01-14,2000-01-15,2000-01-16');

    debug("press enter");
    eventSender.keyDown('\n');
    shouldBeEqualToString('currentMonth()', '2000-01');
    shouldBeEqualToString('selectedValue()', '2000-W02');
    shouldBeEqualToString('selectedDayCells()', '2000-01-10,2000-01-11,2000-01-12,2000-01-13,2000-01-14,2000-01-15,2000-01-16');
    shouldBeEqualToString('highlightedValue()', '2000-W02');
    shouldBeEqualToString('highlightedDayCells()', '2000-01-10,2000-01-11,2000-01-12,2000-01-13,2000-01-14,2000-01-15,2000-01-16');
    shouldBe('document.getElementById("week").value', '"2000-W02"');

    // Test twice in case midnight.
    shouldBeTrue('testToday() || testToday()');

    debug('Check if tabbing works.');
    shouldBe('focusedElement()', '".list-view calendar-table-view"');
    eventSender.keyDown('\t', ['shiftKey']);
    shouldBe('focusedElement()', '".calendar-navigation-button"');
    eventSender.keyDown('\t', ['shiftKey']);
    shouldBe('focusedElement()', '".calendar-navigation-button today-button"');
    eventSender.keyDown('\t', ['shiftKey']);
    shouldBe('focusedElement()', '".calendar-navigation-button"');
    eventSender.keyDown('\t', ['shiftKey']);
    shouldBe('focusedElement()', '".month-popup-button"');

    debug("press esc");
    eventSender.keyDown('\x1B');
    waitUntilClosing(test2AfterClosing);
}

function test2AfterClosing() {
    shouldBeNull('document.getElementById("mock-page-popup")');

    finishJSTest();
}

function testToday() {
    eventSender.keyDown('t');
    var now = new Date();
    var expectedWeek = popupWindow.Week.createFromToday().toString();
    return selectedValue() === expectedWeek && highlightedValue() === expectedWeek;
}

function focusedElement() {
    var element = popupWindow.document.activeElement;
    if (!element)
        return null;
    var identifier = "." + element.className;
    if (element.value)
        identifier += "[value=" + element.value + "]";
    return identifier;
}

</script>
<script src="../../../../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
