LayoutTests/platform/chromium/fast/forms/calendar-picker/week-picker-key-operations-expected.txt

Tests if month picker key bindings work as expected.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Check that page popup doesn't exist at first.
PASS document.getElementById("mock-page-popup") is null
Check that page popup exists.
PASS popupWindow.pagePopupController.toString() is "[object PagePopupController]"
PASS currentMonth() is "2000-01"
PASS selectedValue() is "2000-W01"
PASS selectedDayCells() is "2000-01-03,2000-01-04,2000-01-05,2000-01-06,2000-01-07,2000-01-08,2000-01-09"
PASS highlightedValue() is "2000-W01"
PASS highlightedDayCells() is "2000-01-03,2000-01-04,2000-01-05,2000-01-06,2000-01-07,2000-01-08,2000-01-09"
Check that arrow keys work properly.
press up
PASS currentMonth() is "2000-01"
PASS selectedValue() is "2000-W01"
PASS selectedDayCells() is "2000-01-03,2000-01-04,2000-01-05,2000-01-06,2000-01-07,2000-01-08,2000-01-09"
PASS highlightedValue() is "1999-W52"
PASS highlightedDayCells() is "1999-12-27,1999-12-28,1999-12-29,1999-12-30,1999-12-31,2000-01-01,2000-01-02"
press up
PASS currentMonth() is "1999-12"
PASS selectedValue() is "2000-W01"
PASS selectedDayCells() is ""
PASS highlightedValue() is "1999-W51"
PASS highlightedDayCells() is "1999-12-20,1999-12-21,1999-12-22,1999-12-23,1999-12-24,1999-12-25,1999-12-26"
press down
PASS currentMonth() is "1999-12"
PASS selectedValue() is "2000-W01"
PASS selectedDayCells() is ""
PASS highlightedValue() is "1999-W52"
PASS highlightedDayCells() is "1999-12-27,1999-12-28,1999-12-29,1999-12-30,1999-12-31,2000-01-01"
press down
PASS currentMonth() is "2000-01"
PASS selectedValue() is "2000-W01"
PASS selectedDayCells() is "2000-01-03,2000-01-04,2000-01-05,2000-01-06,2000-01-07,2000-01-08,2000-01-09"
PASS highlightedValue() is "2000-W01"
PASS highlightedDayCells() is "2000-01-03,2000-01-04,2000-01-05,2000-01-06,2000-01-07,2000-01-08,2000-01-09"
press down
PASS currentMonth() is "2000-01"
PASS selectedValue() is "2000-W01"
PASS selectedDayCells() is "2000-01-03,2000-01-04,2000-01-05,2000-01-06,2000-01-07,2000-01-08,2000-01-09"
PASS highlightedValue() is "2000-W02"
PASS highlightedDayCells() is "2000-01-10,2000-01-11,2000-01-12,2000-01-13,2000-01-14,2000-01-15,2000-01-16"
press left
PASS currentMonth() is "2000-01"
PASS selectedValue() is "2000-W01"
PASS selectedDayCells() is "2000-01-03,2000-01-04,2000-01-05,2000-01-06,2000-01-07,2000-01-08,2000-01-09"
PASS highlightedValue() is "2000-W01"
PASS highlightedDayCells() is "2000-01-03,2000-01-04,2000-01-05,2000-01-06,2000-01-07,2000-01-08,2000-01-09"
press right
PASS currentMonth() is "2000-01"
PASS selectedValue() is "2000-W01"
PASS selectedDayCells() is "2000-01-03,2000-01-04,2000-01-05,2000-01-06,2000-01-07,2000-01-08,2000-01-09"
PASS highlightedValue() is "2000-W02"
PASS highlightedDayCells() is "2000-01-10,2000-01-11,2000-01-12,2000-01-13,2000-01-14,2000-01-15,2000-01-16"
press m
PASS currentMonth() is "2000-02"
PASS selectedValue() is "2000-W01"
PASS selectedDayCells() is ""
PASS highlightedValue() is "2000-W07"
PASS highlightedDayCells() is "2000-02-14,2000-02-15,2000-02-16,2000-02-17,2000-02-18,2000-02-19,2000-02-20"
press shift + m
PASS currentMonth() is "2000-01"
PASS selectedValue() is "2000-W01"
PASS selectedDayCells() is "2000-01-03,2000-01-04,2000-01-05,2000-01-06,2000-01-07,2000-01-08,2000-01-09"
PASS highlightedValue() is "2000-W02"
PASS highlightedDayCells() is "2000-01-10,2000-01-11,2000-01-12,2000-01-13,2000-01-14,2000-01-15,2000-01-16"
press y
PASS currentMonth() is "2001-01"
PASS selectedValue() is "2000-W01"
PASS selectedDayCells() is ""
PASS highlightedValue() is "2001-W03"
PASS highlightedDayCells() is "2001-01-15,2001-01-16,2001-01-17,2001-01-18,2001-01-19,2001-01-20,2001-01-21"
press shift + y
PASS currentMonth() is "2000-01"
PASS selectedValue() is "2000-W01"
PASS selectedDayCells() is "2000-01-03,2000-01-04,2000-01-05,2000-01-06,2000-01-07,2000-01-08,2000-01-09"
PASS highlightedValue() is "2000-W02"
PASS highlightedDayCells() is "2000-01-10,2000-01-11,2000-01-12,2000-01-13,2000-01-14,2000-01-15,2000-01-16"
press d
PASS currentMonth() is "2010-01"
PASS selectedValue() is "2000-W01"
PASS selectedDayCells() is ""
PASS highlightedValue() is "2010-W02"
PASS highlightedDayCells() is "2010-01-11,2010-01-12,2010-01-13,2010-01-14,2010-01-15,2010-01-16,2010-01-17"
press shift + d
PASS currentMonth() is "2000-01"
PASS selectedValue() is "2000-W01"
PASS selectedDayCells() is "2000-01-03,2000-01-04,2000-01-05,2000-01-06,2000-01-07,2000-01-08,2000-01-09"
PASS highlightedValue() is "2000-W02"
PASS highlightedDayCells() is "2000-01-10,2000-01-11,2000-01-12,2000-01-13,2000-01-14,2000-01-15,2000-01-16"
press enter
PASS currentMonth() is "2000-01"
PASS selectedValue() is "2000-W02"
PASS selectedDayCells() is "2000-01-10,2000-01-11,2000-01-12,2000-01-13,2000-01-14,2000-01-15,2000-01-16"
PASS highlightedValue() is "2000-W02"
PASS highlightedDayCells() is "2000-01-10,2000-01-11,2000-01-12,2000-01-13,2000-01-14,2000-01-15,2000-01-16"
PASS document.getElementById("week").value is "2000-W02"
PASS testToday() || testToday() is true
Check if tabbing works.
PASS focusedElement() is ".list-view calendar-table-view"
PASS focusedElement() is ".calendar-navigation-button"
PASS focusedElement() is ".calendar-navigation-button today-button"
PASS focusedElement() is ".calendar-navigation-button"
PASS focusedElement() is ".month-popup-button"
press esc
PASS document.getElementById("mock-page-popup") is null
PASS successfullyParsed is true

TEST COMPLETE
